pyproject.toml
[project]
name = "newspaper"
version = "0.1.0"
description = "A CLI tool"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "reportlab>=4.4.9",
]

[project.scripts]
newspaper = "newspaper.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/newspaper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.14",
]
